requirements.txt
absl-py==1.4.0
aiohttp==3.8.4
aiosignal==1.3.1
antlr4-python3-runtime==4.9.3
anyio==3.6.2
appdirs==1.4.4
appnope==0.1.3
argilla==0.0.1
asttokens==2.2.1
async-timeout==4.0.2
attrs==23.1.0
backcall==0.2.0
backoff==2.2.1
black==21.4b2
blinker==1.6.2
cachetools==5.3.0
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.1.0
click==8.1.3
cloudpickle==2.2.1
cohere==4.3.0
comm==0.1.3
contourpy==1.0.7
cryptography==40.0.2
cycler==0.11.0
dataclasses-json==0.5.7
debugpy==1.6.7
decorator==5.1.1
detectron2==0.6
effdet==0.3.0
et-xmlfile==1.1.0
executing==1.2.0
fastapi==0.95.1
filelock==3.12.0
Flask==2.3.1
Flask-Cors==3.0.10
fonttools==4.39.3
frozenlist==1.3.3
fsspec==2023.4.0
future==0.18.3
fvcore==0.1.5.post20221221
google-auth==2.17.3
google-auth-oauthlib==1.0.0
greenlet==2.0.2
grpcio==1.54.0
grpcio-tools==1.54.0
gunicorn==20.1.0
h11==0.14.0
h2==4.1.0
hpack==4.0.0
httpcore==0.17.0
httpx==0.24.0
huggingface-hub==0.14.1
hydra-core==1.3.2
hyperframe==6.0.1
idna==3.4
iopath==0.1.9
ipykernel==6.22.0
ipython==8.13.1
itsdangerous==2.1.2
jedi==0.18.2
Jinja2==3.1.2
joblib==1.2.0
jupyter_client==8.2.0
jupyter_core==5.3.0
kiwisolver==1.4.4
langchain==0.0.153
layoutparser==0.3.4
lxml==4.9.2
Markdown==3.4.3
MarkupSafe==2.1.2
marshmallow==3.19.0
marshmallow-enum==1.5.1
matplotlib==3.7.1
matplotlib-inline==0.1.6
mpmath==1.3.0
msg-parser==1.2.0
multidict==6.0.4
mypy-extensions==1.0.0
nest-asyncio==1.5.6
networkx==3.1
nltk==3.8.1
numexpr==2.8.4
numpy==1.24.3
oauthlib==3.2.2
olefile==0.46
omegaconf==2.3.0
openai==0.27.5
openapi-schema-pydantic==1.2.4
opencv-python==4.6.0.66
openpyxl==3.1.2
packaging==23.1
pandas==2.0.1
parso==0.8.3
pathspec==0.11.1
pdf2image==1.16.3
pdfminer.six==20221105
pdfplumber==0.9.0
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.5.0
platformdirs==3.5.0
portalocker==2.7.0
prompt-toolkit==3.0.38
protobuf==4.22.3
psutil==5.9.5
ptyprocess==0.7.0
pure-eval==0.2.2
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycocotools==2.0.6
pycparser==2.21
pydantic==1.10.7
pydot==1.4.2
Pygments==2.15.1
pypandoc==1.11
pyparsing==3.0.9
pypdf==3.8.1
pytesseract==0.3.10
python-dateutil==2.8.2
python-docx==0.8.11
python-dotenv==1.0.0
python-magic==0.4.27
python-multipart==0.0.6
python-pptx==0.6.21
pytz==2023.3
PyYAML==6.0
pyzmq==25.0.2
qdrant-client==1.1.6
regex==2023.3.23
requests==2.29.0
requests-oauthlib==1.3.1
rsa==4.9
scipy==1.10.1
six==1.16.0
sniffio==1.3.0
SQLAlchemy==2.0.11
stack-data==0.6.2
starlette==0.26.1
sympy==1.11.1
tabulate==0.9.0
tenacity==8.2.2
tensorboard==2.12.2
tensorboard-data-server==0.7.0
tensorboard-plugin-wit==1.8.1
termcolor==2.3.0
timm==0.6.13
toml==0.10.2
torch==2.0.0
torchvision==0.15.1
tornado==6.3.1
tqdm==4.65.0
traitlets==5.9.0
typing-inspect==0.8.0
typing_extensions==4.5.0
tzdata==2023.3
unstructured==0.5.9
unstructured-inference==0.2.4
urllib3==1.26.15
uvicorn==0.22.0
Wand==0.6.11
wcwidth==0.2.6
Werkzeug==2.3.2
XlsxWriter==3.1.0
yacs==0.1.8
yarl==1.9.2
